Fix null varaible in macro.
It seems that if passing an emptry string to a macro and then passing that
variable to another macro will result in the second macro thinking that the
value is not set. This will cause the compiler to fail. To get around this we
stringify the variable and compare it an empty string before deciding how to
call the next macro.