Make Scripts and Images directories optional.
In some cases (particularly with an API) websites that should be deployed as
MVC4 but which are missing certain (optional) directories will cause the NSIS
scripts to fail to compile because the existance of those directories is
required. This change allows for some directories to be missing from the
source folder without causing such errors.